Use useColorModeValue() backport, DRY.
