+++ src/components/RestaurantBannerOverlay.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
-import { useColorModeValue } from 'native-base';
+import { useColorModeValue } from '../styles/theme';
 import { Text } from '@/components/ui/text';
 import { IconClock, IconPlugX } from '@tabler/icons-react-native';
 import i18n from '../i18n';
