+++ src/components/RestaurantMenuHeader.tsx
@@ -1,4 +1,4 @@
-import { useColorModeValue } from 'native-base';
+import { useColorModeValue } from '../styles/theme';
 import { Skeleton, SkeletonText } from '@/components/ui/skeleton';
 import { HStack } from '@/components/ui/hstack';
 import { VStack } from '@/components/ui/vstack';
