+++ src/components/RestaurantProductBadge.tsx
@@ -4,7 +4,7 @@ import {
   IconRecycle,
 } from '@tabler/icons-react-native';
 import i18next from 'i18next';
-import { useColorModeValue } from 'native-base';
+import { useColorModeValue } from '../styles/theme';
 import { Text } from '@/components/ui/text';
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
